Split GlobalStoragePool
